requirements.txt
# Core CLI
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0

# Database
sqlalchemy>=2.0.35

# Browser Automation
playwright>=1.49.0
playwright-stealth>=1.0.6

# AI
openai>=1.58.0

# Scheduling
apscheduler==3.10.4

# Security
cryptography==41.0.7

# Utilities
loguru==0.7.2
faker==22.0.0
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
